requirements.txt
streamlit>=1.35.0
streamlit-folium>=0.20.0
folium>=0.17.0
ultralytics>=8.2.0
Pillow>=10.0.0
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
numpy>=1.26.0
streamlit-drawable-canvas>=0.9.3
